fix reset config

+++ consts.py
@@ -16,18 +16,27 @@ MAIN_FONT = os.path.join(CONFIG_DIR, 'JetBrainsMonoNerdFont-Regular.ttf')
 
 with open(BASE_CONFIG, 'r') as file:
     BASE_SETTINGS = yaml.load(file, Loader=yaml.SafeLoader)
-# Replace FONT_PATH with real path
-with open(CONFIG_FILE, 'r') as file:
-    content = file.read()
-new_content = content.replace("FONT_PATH", MAIN_FONT)
-with open(CONFIG_FILE, 'w') as file:
-    file.write(new_content)
+
+def replace_placeholder_path():
+    if os.path.isfile(CONFIG_FILE):
+        with open(CONFIG_FILE, 'r') as file:
+            content = file.read()
+        new_content = content.replace("PLACEHOLDER_FONT_PATH", MAIN_FONT)
+        with open(CONFIG_FILE, 'w') as file:
+            file.write(new_content)
+
+def make_base_config():
+    os.makedirs(CONFIG_DIR, exist_ok=True)  # safely create directory if needed
+    shutil.copyfile(BASE_CONFIG, CONFIG_FILE)
+    replace_placeholder_path()
+    shutil.copyfile(BASE_FONT, MAIN_FONT)
 
 if not os.path.isfile(CONFIG_FILE):
     print(f"{CONFIG_FILE} not found. Making basic config...")
-    os.makedirs(CONFIG_DIR, exist_ok=True)  # safely create directory if needed
-    shutil.copyfile(BASE_CONFIG, CONFIG_FILE)
-    shutil.copyfile(BASE_FONT, MAIN_FONT)
+    make_base_config()
+
+# Replace PLACEHOLDER_FONT_PATH with real path
+replace_placeholder_path()
 
 with open(CONFIG_FILE, 'r') as file:
     SETTINGS = yaml.load(file, Loader=yaml.SafeLoader)
@@ -39,8 +48,7 @@ for setting in BASE_SETTINGS:
             ans = input(f"{setting} not found in config file, reset to default config (y/n)? ")
         if 'y' in ans.lower():
             print("Resetting config...")
-            shutil.copyfile(BASE_CONFIG, CONFIG_FILE)
-            shutil.copyfile(BASE_FONT, MAIN_FONT)
+            make_base_config()
             break
 with open(CONFIG_FILE, 'r') as file:
     SETTINGS = yaml.load(file, Loader=yaml.SafeLoader)
